# samples/express/docs/openapi.yaml
...
openapi: 3.1.0
info:
  title: Perxona Connect API
  description: The Perxona Connect API — authentication, avatar catalog, and TTS token routes for the sv-presenter Web Component.
  version: 0.1.0
servers: []
paths:
  /api/v1/connect/auth/login:
    post:
      tags:
        - connect
        - connect
      summary: Authenticate a Connect user and issue an access token
      description: >-
        Authenticate with email/password and receive a JWT access token signed with the dedicated Connect secret. Pass
        the token as `Authorization: Bearer <token>` on subsequent Connect API calls.


        ## Rate Limit 

        This endpoint is limit __5 calls / 1 seconds__
      operationId: post_v1_connect_connect_login
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConnectLoginRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectLoginResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/connect/assets/avatars:
    get:
      tags:
        - connect
        - connect
      summary: List avatars
      description: Return a paginated list of active avatars accessible to the authenticated user's organization.
      operationId: get_v1_connect_list_avatars
      security:
        - HTTPBearer: []
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
            title: Page
        - name: size
          in: query
          required: false
          schema:
            type: integer
            maximum: 100
            minimum: 1
            default: 50
            title: Size
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_ConnectListAvatarResponse_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/connect/assets/avatars/{avatar_id}:
    get:
      tags:
        - connect
        - connect
      summary: Get avatar detail
      description: Return full detail for a single active avatar including LOD URLs and lipsync support.
      operationId: get_v1_connect_get_avatar
      security:
        - HTTPBearer: []
      parameters:
        - name: avatar_id
          in: path
          required: true
          schema:
            type: string
            title: Avatar Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectAvatarDetailResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/connect/assets/avatars/{avatar_id}/motions:
    get:
      tags:
        - connect
        - connect
      summary: List avatar motions
      description: >-
        Return a paginated list of active motions compatible with the avatar's skeleton, optionally filtered by pose
        tag.
      operationId: get_v1_connect_list_connect_avatar_motions
      security:
        - HTTPBearer: []
      parameters:
        - name: avatar_id
          in: path
          required: true
          schema:
            type: string
            title: Avatar Id
        - name: pose_tag
          in: query
          required: false
          schema:
            anyOf:
              - type: string
                pattern: ^pose:[a-z0-9_]+$
              - type: 'null'
            title: Pose Tag
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
            title: Page
        - name: size
          in: query
          required: false
          schema:
            type: integer
            maximum: 100
            minimum: 1
            default: 50
            title: Size
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_ConnectMotionAssetResponse_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/connect/assets/scenes:
    get:
      tags:
        - connect
        - connect
      summary: List scenes
      description: Return a paginated list of active scenes accessible to the authenticated user's organization.
      operationId: get_v1_connect_list_scenes
      security:
        - HTTPBearer: []
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
            title: Page
        - name: size
          in: query
          required: false
          schema:
            type: integer
            maximum: 100
            minimum: 1
            default: 50
            title: Size
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_ConnectListSceneResponse_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/connect/assets/scenes/{scene_id}:
    get:
      tags:
        - connect
        - connect
      summary: Get scene detail
      description: Return full detail for a single active scene including LOD URLs.
      operationId: get_v1_connect_get_scene
      security:
        - HTTPBearer: []
      parameters:
        - name: scene_id
          in: path
          required: true
          schema:
            type: string
            title: Scene Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectSceneDetailResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/connect/voices:
    get:
      tags:
        - connect
        - connect
      summary: List available custom voice models
      description: >-
        Return a paginated list of custom voices available to the organization. Supports filtering by language and
        provider.


        ## Rate Limit 

        This endpoint is limit __30 calls / 1 seconds__
      operationId: get_v1_connect_connect_list_voices
      security:
        - HTTPBearer: []
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
            title: Page
        - name: size
          in: query
          required: false
          schema:
            type: integer
            maximum: 100
            minimum: 1
            default: 50
            title: Size
        - name: search_term
          in: query
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Search Term
        - name: date_range_type
          in: query
          required: false
          schema:
            anyOf:
              - $ref: '#/components/schemas/DateRangeStrEnum'
              - type: 'null'
            title: Date Range Type
        - name: start_at
          in: query
          required: false
          schema:
            anyOf:
              - type: integer
              - type: 'null'
            title: Start At
        - name: end_at
          in: query
          required: false
          schema:
            anyOf:
              - type: integer
              - type: 'null'
            title: End At
        - name: language
          in: query
          required: false
          schema:
            anyOf:
              - $ref: '#/components/schemas/Language'
              - type: 'null'
            title: Language
        - name: provider
          in: query
          required: false
          schema:
            anyOf:
              - $ref: '#/components/schemas/VoiceProvider'
              - type: 'null'
            title: Provider
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_ConnectVoiceResponse_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    AssetLOD:
      type: string
      enum:
        - lod0
        - lod1
        - lod2
      title: AssetLOD
      description: The level of detail for the asset
...
    AvatarThumbnailType:
      type: string
      enum:
        - default
        - head
        - upper_body
        - full_body
        - three_quarter_face
        - animated_full_body
        - hd
        - loading
      title: AvatarThumbnailType
      description: The type of thumbnail for the asset
...
    ConnectAvatarDetailResponse:
      properties:
        avatar_id:
          type: string
          title: Avatar Id
        name:
          type: string
          title: Name
        description:
          type: string
          title: Description
        thumbnail_urls:
          additionalProperties:
            type: string
          propertyNames:
            $ref: '#/components/schemas/AvatarThumbnailType'
          type: object
          title: Thumbnail Urls
        tags:
          items:
            type: string
          type: array
          title: Tags
        lod_urls:
          additionalProperties:
            type: string
          propertyNames:
            $ref: '#/components/schemas/AssetLOD'
          type: object
          title: Lod Urls
        lipsync_configs:
          additionalProperties:
            additionalProperties:
              $ref: '#/components/schemas/LipSyncConfig-Output'
            type: object
          propertyNames:
            $ref: '#/components/schemas/LipSyncMode'
          type: object
          title: Lipsync Configs
      type: object
      required:
        - avatar_id
        - name
        - description
        - thumbnail_urls
        - tags
        - lod_urls
        - lipsync_configs
      title: ConnectAvatarDetailResponse
...
    ConnectListAvatarResponse:
      properties:
        avatar_id:
          type: string
          title: Avatar Id
        name:
          type: string
          title: Name
        description:
          type: string
          title: Description
        thumbnail_urls:
          additionalProperties:
            type: string
          propertyNames:
            $ref: '#/components/schemas/AvatarThumbnailType'
          type: object
          title: Thumbnail Urls
        tags:
          items:
            type: string
          type: array
          title: Tags
      type: object
      required:
        - avatar_id
        - name
        - description
        - thumbnail_urls
        - tags
      title: ConnectListAvatarResponse
...
    ConnectLoginRequest:
      properties:
        email:
          type: string
          title: Email
        password:
          type: string
          title: Password
      type: object
      required:
        - email
        - password
      title: ConnectLoginRequest
      description: The Connect login request schema
    ConnectLoginResponse:
      properties:
        access_token:
          type: string
          title: Access Token
      type: object
      required:
        - access_token
      title: ConnectLoginResponse
      description: The Connect login response carrying the access token only
    ConnectMotionAssetResponse:
      properties:
        motion_id:
          type: string
          title: Motion Id
        name:
          type: string
          title: Name
        tags:
          items:
            type: string
          type: array
          title: Tags
        lod_urls:
          additionalProperties:
            type: string
          propertyNames:
            $ref: '#/components/schemas/AssetLOD'
          type: object
          title: Lod Urls
      type: object
      required:
        - motion_id
        - name
        - tags
        - lod_urls
      title: ConnectMotionAssetResponse
...
    ConnectSceneDetailResponse:
      properties:
        scene_id:
          type: string
          title: Scene Id
        name:
          type: string
          title: Name
        description:
          type: string
          title: Description
        thumbnail_urls:
          additionalProperties:
            type: string
          propertyNames:
            $ref: '#/components/schemas/SceneThumbnailType'
          type: object
          title: Thumbnail Urls
        tags:
          items:
            type: string
          type: array
          title: Tags
        lod_urls:
          additionalProperties:
            type: string
          propertyNames:
            $ref: '#/components/schemas/AssetLOD'
          type: object
          title: Lod Urls
      type: object
      required:
        - scene_id
        - name
        - description
        - thumbnail_urls
        - tags
        - lod_urls
      title: ConnectSceneDetailResponse
...
    DateRangeStrEnum:
      type: string
      enum:
        - past_7_days
        - past_14_days
        - past_30_days
        - custom
      title: DateRangeStrEnum
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    Language:
      type: string
      enum:
        - auto
        - ar
        - bn
        - bg
        - hr
        - cs
        - da
        - nl
        - en
        - et
        - fi
        - fr
        - de
        - el
        - he
        - hi
        - hu
        - id
        - it
        - ja
        - ko
        - lv
        - lt
        - nb
        - pl
        - pt
        - ro
        - ru
        - sr
        - sk
        - sl
        - es
        - sw
        - sv
        - th
        - tr
        - uk
        - vi
        - zh
      title: Language
    LipSyncConfig-Output:
      properties:
        config:
          anyOf:
            - type: string
            - type: 'null'
          title: Config
        resource:
          anyOf:
            - type: string
            - type: 'null'
          title: Resource
        config_url:
          anyOf:
            - type: string
            - type: 'null'
          title: Config Url
          readOnly: true
        resource_url:
          anyOf:
            - type: string
            - type: 'null'
          title: Resource Url
          readOnly: true
      type: object
      required:
        - config_url
        - resource_url
      title: LipSyncConfig
      description: The lip sync configuration
    LipSyncMode:
      type: string
      enum:
        - wlipsync
        - xrlipsync
      title: LipSyncMode
      description: The lip sync mode
    Page_ConnectListAvatarResponse_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/ConnectListAvatarResponse'
          type: array
          title: Items
        total:
          type: integer
          minimum: 0
          title: Total
        page:
          type: integer
          minimum: 1
          title: Page
        size:
          type: integer
          minimum: 1
          title: Size
        pages:
          type: integer
          minimum: 0
          title: Pages
      type: object
      required:
        - items
        - total
        - page
        - size
        - pages
      title: Page[ConnectListAvatarResponse]
    Page_ConnectListSceneResponse_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/ConnectListSceneResponse'
          type: array
          title: Items
        total:
          type: integer
          minimum: 0
          title: Total
        page:
          type: integer
          minimum: 1
          title: Page
        size:
          type: integer
          minimum: 1
          title: Size
        pages:
          type: integer
          minimum: 0
          title: Pages
      type: object
      required:
        - items
        - total
        - page
        - size
        - pages
      title: Page[ConnectListSceneResponse]
    Page_ConnectMotionAssetResponse_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/ConnectMotionAssetResponse'
          type: array
          title: Items
        total:
          type: integer
          minimum: 0
          title: Total
        page:
          type: integer
          minimum: 1
          title: Page
        size:
          type: integer
          minimum: 1
          title: Size
        pages:
          type: integer
          minimum: 0
          title: Pages
      type: object
      required:
        - items
        - total
        - page
        - size
        - pages
      title: Page[ConnectMotionAssetResponse]
    Page_ConnectVoiceResponse_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/ConnectVoiceResponse'
          type: array
          title: Items
        total:
          type: integer
          minimum: 0
          title: Total
        page:
          type: integer
          minimum: 1
          title: Page
        size:
          type: integer
          minimum: 1
          title: Size
        pages:
          type: integer
          minimum: 0
          title: Pages
      type: object
      required:
        - items
        - total
        - page
        - size
        - pages
      title: Page[ConnectVoiceResponse]
    SceneThumbnailType:
      type: string
      enum:
        - default
        - high_res
        - loading
      title: SceneThumbnailType
      description: The type of thumbnail for a scene asset
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
              - type: string
              - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
        - loc
        - msg
        - type
      title: ValidationError
    VoiceProvider:
      type: string
      enum:
        - google
        - aws
        - azure
        - elevenlabs
      title: VoiceProvider
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer